Refactor build-data
